+++ .infra/playbooks/roles/setup/files/app/tools/mongodb/backup-mongodb.sh
@@ -9,9 +9,9 @@ readonly BACKUP_FILE="${BACKUP_LOCAL_DIR}/mongodb-$(date +'%Y-%m-%d_%H%M%S').gpg
 function backup() {
   echo "Creating backup..."
   mkdir -p "${BACKUP_LOCAL_DIR}"
-  docker run --rm -i mongo:6.0.2-focal mongodump --archive --gzip --uri="$REPLY" \
+  docker run --rm -i mongo:6.0.2-focal mongodump --archive --gzip --uri="{{vault[env_type].MNA_BAL_MONGODB_BACKUP_URI}}" \
     | bash "${SCRIPT_DIR}/../gpg/encrypt.sh" > "${BACKUP_FILE}"
-  rm /opt/bal/backups/mongodb/latest.gpg
+  rm -f /opt/bal/backups/mongodb/latest.gpg
   ln -s "${BACKUP_FILE}" /opt/bal/backups/mongodb/latest.gpg
 }